Minor package and visibility refactors
